Unprintable characters in section names are now escaped (fixes #3). The PE's size is now cached and won't be computed everytime get_filesize is called. This prevents useless I/O (#4).

+++ plugins/plugin_resources.cpp
@@ -91,7 +91,7 @@ public:
 			}
 		}
 
-		float ratio = (float) size / (float) pe.get_filesize();
+		double ratio = static_cast<double>(size) / static_cast<double>(pe.get_filesize());
 		if (ratio > .75)
 		{
 			std::stringstream ss;
